Rename client to better match triton client
